refactor: FSD 패턴에 맞도록 query 분리
